revert: ALSA: hda - Distinguish each substream for better sticky assignment
reverted commit: d5cf9911988287e819ce98ccd9f61ca82fbc90c6

"
The commit ef18beded8ddbaafdf4914bab209f77e60ae3a18 introduced a
mechanism to assign the previously used slot for the next reopen of a
PCM stream.  But the PCM device number isn't always unique (it may
have multiple substreams), and also the code doesn't check the stream
direction, thus both playback and capture streams share the same
device number.


For avoiding this conflict, make a unique key for each substream and
store/check this value at reopening.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
"

This commit introduce a regression since it assign a wrong
subdevice to the substream.

On first open for a subdevice the condition
if (res->assigned_key == key)
Can't match anything and finish to assign the last device in the list.
The break is never reached.

On our robot it finish to get first substream assigned to last subdevice,
the second substream is assigned to the first.

The multi stream is out of sync between these two subdevice,
and introduce a late of 6 samples on the two first channel.