ath6kl: Fix WLAN tethering authentication problem.
When enabling WLAN tethering with connman, a new AP was visible and a
STA could connect to it. When the STA tried to authenticate to the newly
created AP, the WPA authentication mechanism was stuck in the 1/4 msg of
4-Way Handshake.

In ath6kl_rx(), the ath6kl_find_sta() function is looking for the
h_source field of the Ethernet frame header received by the STA. The
datap pointer that points to the Ethernet frame header was incorrect,
and was pointing at the wrong offset in the buffer.

This commit adds a pad_before_data_start offset to set the datap pointer
to the Ethernet frame header. datap->h_source parameter is now really
pointing to the source ethernet address and the authentication process
can happen.

Related to #34880

Change-Id: I3f1bf614a67f2f23804c897c1b93960445db6923
Reviewed-on: http://gerrit.aldebaran.lan/72821
Tested-by: plemagourou <plemagourou@aldebaran.com>
Reviewed-by: nrubinstein <nrubinstein@aldebaran-robotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/81931
Reviewed-by: jmassot <jmassot@aldebaran-robotics.com>