ath6kl: Fix Wifi connection disconnect after 1 min.
One minute after a successful connection, the kernel checks if the
frequency and the channel width are well configured for the country we
are in. (regulatory domain)

ath6kl driver was setting the NL80211_CHAN_HT20 channel without checking
for the HT capabilities. (we should have NL80211_CHAN_NO_HT in our
case because the firmware did not support HT)

This patch adds a check on ht_cap.ht_supported in order to create the
channel corresponding to the firmware capabilities.

Related to #34880

Change-Id: Ie94cb4b58647c04bf0d1df08ca48db401d6d8af2
Reviewed-on: http://gerrit.aldebaran.lan/72822
Tested-by: plemagourou <plemagourou@aldebaran.com>
Reviewed-by: nrubinstein <nrubinstein@aldebaran-robotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/81932
Reviewed-by: jmassot <jmassot@aldebaran-robotics.com>