ath6kl: Unset IFF_LOWER_UP flags when stopping AP mode.
When disabling tethering with connman after having enabled it, the
ath6kl driver was not setting the IFF_LOWER_UP flag to 0 (carrier off)
on the wlan interface.

Hence, connman was not calling handle_tethering() and tethering was not
activated on the next tethering activation call (connmanctl tether wifi on).

Add a netif_carrier_off() call when stopping AP mode fixes the problem.

Fix #34880

Change-Id: Ic664361b012f39554aca2661c8b6317dce6dbd47
Reviewed-on: http://gerrit.aldebaran.lan/72823
Tested-by: plemagourou <plemagourou@aldebaran.com>
Reviewed-by: nrubinstein <nrubinstein@aldebaran-robotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/81933
Reviewed-by: jmassot <jmassot@aldebaran-robotics.com>